Revert "Making timeout a long as newly generated annotations have it (we pulled the generation code from a recent commit, needed this code to reflect it too)"
This reverts commit 8324c738440249e92772d6e667409b7fafd4f74d.

+++ android/src/cucumber/runtime/android/JavaHookDefinition.java
@@ -18,12 +18,12 @@ import cucumber.runtime.java.ObjectFactory;
 class JavaHookDefinition implements HookDefinition {
 
     private final Method method;
-    private final long timeout;
+    private final int timeout;
     private final TagExpression tagExpression;
     private final int order;
     private final ObjectFactory objectFactory;
 
-    public JavaHookDefinition(Method method, String[] tagExpressions, int order, long timeout, ObjectFactory objectFactory) {
+    public JavaHookDefinition(Method method, String[] tagExpressions, int order, int timeout, ObjectFactory objectFactory) {
         this.method = method;
         this.timeout = timeout;
         tagExpression = new TagExpression(asList(tagExpressions));
