+++ java/src/main/java/cucumber/runtime/java/JavaBackend.java
@@ -1,12 +1,5 @@
 package cucumber.runtime.java;
 
-import gherkin.formatter.model.Step;
-
-import java.lang.annotation.Annotation;
-import java.lang.reflect.Method;
-import java.util.List;
-import java.util.regex.Pattern;
-
 import cucumber.api.java.After;
 import cucumber.api.java.Before;
 import cucumber.runtime.Backend;
@@ -18,6 +11,12 @@ import cucumber.runtime.Utils;
 import cucumber.runtime.io.ClasspathResourceLoader;
 import cucumber.runtime.io.ResourceLoader;
 import cucumber.runtime.snippets.SnippetGenerator;
+import gherkin.formatter.model.Step;
+
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Method;
+import java.util.List;
+import java.util.regex.Pattern;
 
 public class JavaBackend implements Backend {
     private final SnippetGenerator snippetGenerator = new SnippetGenerator(new JavaSnippet());
@@ -104,9 +103,9 @@ public class JavaBackend implements Backend {
         return Pattern.compile(regexpString);
     }
 
-    private long timeout(Annotation annotation) throws Throwable {
+    private int timeout(Annotation annotation) throws Throwable {
         Method regexpMethod = annotation.getClass().getMethod("timeout");
-        return (Long) Utils.invoke(annotation, regexpMethod, 0);
+        return (Integer) Utils.invoke(annotation, regexpMethod, 0);
     }
 
     void addHook(Annotation annotation, Method method) {
@@ -114,11 +113,11 @@ public class JavaBackend implements Backend {
 
         if (annotation.annotationType().equals(Before.class)) {
             String[] tagExpressions = ((Before) annotation).value();
-            long timeout = ((Before) annotation).timeout();
+            int timeout = ((Before) annotation).timeout();
             glue.addBeforeHook(new JavaHookDefinition(method, tagExpressions, ((Before) annotation).order(), timeout, objectFactory));
         } else {
             String[] tagExpressions = ((After) annotation).value();
-            long timeout = ((After) annotation).timeout();
+            int timeout = ((After) annotation).timeout();
             glue.addAfterHook(new JavaHookDefinition(method, tagExpressions, ((After) annotation).order(), timeout, objectFactory));
         }
     }
